+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.20.0 - (Spring Green Vienna) - June 2022
 
 - added default value (`aggregates.mdsp.json`) for `mdsp aggregates --download` parameter
+- improved `mdsp aggregates` command. all parameters, including intervalunit and intervalvalue can be specified.
 - Created a new `mdsp timeseries` command to match the `mdsp aggregates` command. [#318](https://github.com/mindsphere/mindconnect-nodejs/issues/318)
 - Bumped most depedencies